requirements.txt
Flask==2.3.3
tensorflow==2.13.0
keras==2.13.1
numpy==1.24.3
opencv-python==4.8.1.78
Pillow==10.0.1
matplotlib==3.7.2
Werkzeug==2.3.7
